remove obsolete comments; move redis service config
