+++ src/ProductConstructionService/ProductConstructionService.Api/PcsStartup.cs
@@ -36,7 +36,6 @@ using ProductConstructionService.Api.Queue;
 using ProductConstructionService.Api.Telemetry;
 using ProductConstructionService.Api.VirtualMonoRepo;
 using ProductConstructionService.Common;
-using StackExchange.Redis;
 
 namespace ProductConstructionService.Api;
 
@@ -108,17 +107,14 @@ internal static class PcsStartup
     /// Registers all necessary services for the Product Construction Service
     /// </summary>
     /// <param name="builder"></param>
-    /// <param name="isDevelopment">Is this development environment?</param>
     /// <param name="addKeyVault">Use KeyVault for secrets?</param>
     /// <param name="addSwagger">Add Swagger UI?</param>
     internal static async Task ConfigurePcs(
-    //internal static void ConfigurePcs(
         this WebApplicationBuilder builder,
-        bool isDevelopment,
         bool addKeyVault,
         bool addSwagger)
     {
-        //Currently this just means cloning the VMR
+        bool isDevelopment = builder.Environment.IsDevelopment();
         bool initializeService = !isDevelopment;
 
         // Read configuration
@@ -150,7 +146,7 @@ internal static class PcsStartup
         builder.Services.AddGitHubTokenProvider();
         builder.Services.AddScoped<IRemoteFactory, DarcRemoteFactory>();
         builder.Services.AddSingleton<Microsoft.Extensions.Internal.ISystemClock, Microsoft.Extensions.Internal.SystemClock>();
-        builder.Services.AddSingleton<Microsoft.DotNet.Services.Utility.ExponentialRetry>();
+        builder.Services.AddSingleton<ExponentialRetry>();
         builder.Services.Configure<ExponentialRetryOptions>(_ => { });
         builder.Services.AddMemoryCache();
         builder.Services.AddSingleton(builder.Configuration);
@@ -163,7 +159,7 @@ internal static class PcsStartup
         builder.Services.AddMergePolicies();
         builder.Services.Configure<SlaOptions>(builder.Configuration.GetSection(ConfigurationKeys.DependencyFlowSLAs));
 
-        await ConfigureRedis(builder, isDevelopment, managedIdentityId);
+        await builder.Services.ConfigureRedis(builder.Configuration, isDevelopment, managedIdentityId);
 
         if (initializeService)
         {
@@ -290,25 +286,5 @@ internal static class PcsStartup
     public static bool IsGet(this HttpContext context)
     {
         return string.Equals(context.Request.Method, "get", StringComparison.OrdinalIgnoreCase);
-    }
-
-    public static async Task ConfigureRedis(WebApplicationBuilder builder, bool isDevelopment, string? managedIdentityId)
-    {
-        var redisConfig = ConfigurationOptions.Parse(
-            builder.Configuration.GetConnectionString("redis") ?? throw new Exception("Missing Redis connection string."));
-
-        // Local redis instance should not need authentication
-        if (!isDevelopment)
-        {
-            var azureOptions = new Microsoft.Azure.StackExchangeRedis.AzureCacheOptions();
-            if (managedIdentityId != null && !managedIdentityId.Equals("system"))
-            {
-                azureOptions.ClientId = managedIdentityId;
-            }
-
-            await redisConfig.ConfigureForAzureAsync(azureOptions);
-        }
-
-        builder.Services.AddSingleton<IConnectionMultiplexer>(_ => ConnectionMultiplexer.Connect(redisConfig));
     }
 }
