+++ src/ProductConstructionService/ProductConstructionService.Api/Program.cs
@@ -16,7 +16,6 @@ bool useSwagger = isDevelopment;
 
 await builder.ConfigurePcs(
     addKeyVault: true,
-    isDevelopment: isDevelopment,
     addSwagger: useSwagger);
 
 var app = builder.Build();
