+++ src/ProductConstructionService/ProductConstructionService.Common/ConfigurationKeys.cs
@@ -1,0 +1,8 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+namespace ProductConstructionService.Common;
+public static class ConfigurationKeys
+{
+    public const string RedisConnectionString = "redis";
+}
