+++ src/ProductConstructionService/ProductConstructionService.Common/ProductConstructionService.Common.csproj
@@ -14,6 +14,8 @@
     <PackageReference Include="Microsoft.Extensions.Configuration" />
     <PackageReference Include="Microsoft.Extensions.Logging.ApplicationInsights" />
     <PackageReference Include="Microsoft.Extensions.Logging.Console" />
+    <PackageReference Include="Microsoft.Azure.StackExchangeRedis" />
+    <PackageReference Include="NRedisStack" />
   </ItemGroup>
 
   <ItemGroup>
