+++ src/ProductConstructionService/ProductConstructionService.Common/ProductConstructionServiceExtension.cs
@@ -13,6 +13,8 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 using Microsoft.EntityFrameworkCore;
+using StackExchange.Redis;
+using Microsoft.Azure.StackExchangeRedis;
 
 namespace ProductConstructionService.Common;
 
@@ -72,5 +74,29 @@ public static class ProductConstructionServiceExtension
 
         services.AddKustoClientProvider("Kusto");
         services.AddSingleton<IInstallationLookup, BuildAssetRegistryInstallationLookup>(); ;
+    }
+
+    public static async Task ConfigureRedis(
+        this IServiceCollection services,
+        IConfiguration configuration,
+        bool isDevelopment,
+        string? managedIdentityId)
+    {
+        var redisConfig = ConfigurationOptions.Parse(
+            configuration.GetSection("ConnectionStrings").GetRequiredValue(ConfigurationKeys.RedisConnectionString));
+
+        // Local redis instance should not need authentication
+        if (!isDevelopment)
+        {
+            AzureCacheOptions azureOptions = new();
+            if (managedIdentityId != null && !managedIdentityId.Equals("system"))
+            {
+                azureOptions.ClientId = managedIdentityId;
+            }
+
+            await redisConfig.ConfigureForAzureAsync(azureOptions);
+        }
+
+        services.AddSingleton<IConnectionMultiplexer>(_ => ConnectionMultiplexer.Connect(redisConfig));
     }
 }
